+++ swov1/src/funcs/dboPluginOperation.ts
@@ -39,6 +39,7 @@ export function dboPluginOperation(
     void,
     | errors.CommonBadRequestErrorResponse
     | errors.CommonUnauthorizedErrorResponse
+    | errors.CommonForbiddenErrorResponse
     | errors.CommonNotFoundErrorResponse
     | errors.CommonInternalErrorResponse
     | SwoError
@@ -68,6 +69,7 @@ async function $do(
       void,
       | errors.CommonBadRequestErrorResponse
       | errors.CommonUnauthorizedErrorResponse
+      | errors.CommonForbiddenErrorResponse
       | errors.CommonNotFoundErrorResponse
       | errors.CommonInternalErrorResponse
       | SwoError
@@ -175,6 +177,7 @@ async function $do(
     void,
     | errors.CommonBadRequestErrorResponse
     | errors.CommonUnauthorizedErrorResponse
+    | errors.CommonForbiddenErrorResponse
     | errors.CommonNotFoundErrorResponse
     | errors.CommonInternalErrorResponse
     | SwoError
@@ -189,6 +192,7 @@ async function $do(
     M.nil(204, z.void()),
     M.jsonErr(400, errors.CommonBadRequestErrorResponse$inboundSchema),
     M.jsonErr(401, errors.CommonUnauthorizedErrorResponse$inboundSchema),
+    M.jsonErr(403, errors.CommonForbiddenErrorResponse$inboundSchema),
     M.jsonErr(404, errors.CommonNotFoundErrorResponse$inboundSchema),
     M.jsonErr(500, errors.CommonInternalErrorResponse$inboundSchema),
     M.fail("4XX"),