+++ swov1/src/lib/config.ts
@@ -71,8 +71,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.16",
-  sdkVersion: "0.11.0",
-  genVersion: "2.884.0",
+  sdkVersion: "0.12.0",
+  genVersion: "2.889.1",
   userAgent:
-    "speakeasy-sdk/typescript 0.11.0 2.884.0 1.0.16 @solarwinds/swo-sdk-typescript",
+    "speakeasy-sdk/typescript 0.12.0 2.889.1 1.0.16 @solarwinds/swo-sdk-typescript",
 } as const;